requirements.txt
pefile==2023.2.7
yara-python==4.3.1
scapy==2.5.0
reportlab==4.1.0
stix2==3.0.1
mitreattack-python==2.1.0
onnxruntime
python-dateutil==2.8.2
